+++ src/test/java/org/plumelib/util/CollectionsPlumeTest.java
@@ -421,6 +421,15 @@ public final class CollectionsPlumeTest {
   }
 
   @Test
+  public void testListFilter() {
+    List<Integer> in = Arrays.asList(new Integer[] {1, 2, 3, 4, 5});
+    List<Integer> odd = Arrays.asList(new Integer[] {1, 3, 5});
+    List<Integer> even = Arrays.asList(new Integer[] {2, 4});
+    assertEquals(odd, CollectionsPlume.filter(in, i -> i % 2 == 1));
+    assertEquals(even, CollectionsPlume.filter(in, i -> i % 2 == 0));
+  }
+
+  @Test
   public void testSortedSetEquals() {
     TreeSet<Integer> s2 = new TreeSet<>(Arrays.asList(1, 2));
     TreeSet<Integer> s3 = new TreeSet<>(Arrays.asList(1, 2, 3));